requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.4.3
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.13.2
urllib3==2.4.0
win32_setctime==1.2.0
yarl==1.19.0
